back/requirements.txt
python-dotenv
flask
flask-cors
boto3